Create README.md
Three peaks Tanzania safari was created with love, and out of decades of experience in safaris, to offer what no other tour operator was able to offer: Unparalleled flexibility and dedication. Our team will always go the extra mile to make your journey unforgettable. Since the first moment you contact us, until it’s time to say goodbye, you will have instant customer care, high-end equipment, amazing vehicles and amazing surprises like our unique and signatures hot lunches in the wild (instead of lunch boxes). But that’s still not the most important aspect of our service…

At Three peaks Tanzania safari, we believe that the guide makes a safari. And we have dedicated all our efforts into building a team of guides that we can proudly say are the best of the best: Extremely professional and experienced but at the same time cheerful, talkative, funny and as flexible as it gets.